Renamed api files to avoid object-file clashes
